+++ CHANGELOG.md
@@ -10,10 +10,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Optional latency histogram report to stdout.
 - Optional full latency CSV report to a given filename.
 - Respect `http_proxy` environment variable.
+- Added `-totalRequests` flag for exiting after the given number of requests are issued.
 ### Changed
 - Removed `-reuse` deprecation warning.
 - Removed bytes received from the output.
 - Removed `-url`, instead use the first argument from ARGV.
+- Use the new `/net/http/httptrace` package and measure latency as time to first byte.
+- No longer exit with an error code of 1 after cleanup
 
 ## [0.7.0] - 2016-07-21
 ### Added